Fixed NotFoundError appearing in browser console
Issue:
On some systems / browsers an error called 'NotFoundError: Node was not found' appears in the developer console. However, this error does not affect the notification functionality at all (just a cosmetic thing).

Solution:
Rewriting the event listener remover for the 'transitionend' event, with values not based on the event object but the notification reference and the event type name itself, solves this problem.