Fixed wrong symbol size and added resource check
Issue:
Using round symbol corners together with an svg icon causes the symbol being to large. Furthermore making the symbol visible but not setting a resource results into an error.

Solution:
The symbol size is now being calculated correctly. When the symbol is set to be visible but the resource is not configured (and the profile is not 'success' or 'error', so no default icon exists), the symbol will not be rendered.